Move common utility methods to utils/common.js.
