+++ packages/ckeditor5-table/src/commands/insertrowcommand.js
@@ -8,8 +8,8 @@
  */
 
 import Command from '@ckeditor/ckeditor5-core/src/command';
-import { findAncestor } from './utils';
 import { getRowIndexes, getSelectionAffectedTableCells } from '../utils/selection';
+import { findAncestor } from '../utils/common';
 
 /**
  * The insert row command.
