+++ packages/ckeditor5-table/src/commands/mergecellcommand.js
@@ -9,8 +9,8 @@
 
 import Command from '@ckeditor/ckeditor5-core/src/command';
 import TableWalker from '../tablewalker';
-import { isHeadingColumnCell, findAncestor } from './utils';
 import { getTableCellsContainingSelection } from '../utils/selection';
+import { findAncestor, isHeadingColumnCell } from '../utils/common';
 
 /**
  * The merge cell command.
